[Web UI] Avoid relative imports with module-resolver plugin (#1363)
Any absolute es6 import statement is resolved relative to the `src` directory. This avoids the need to use deep relative paths that reference a file's current location.

Compared to a set of webpack aliases to specific directories within the project, this approach is more flexible, less fragile, and easier to reason.

In order to use `eslint-import-resolver-babel-module`, `babel-preset-react-app` had to be abandoned. The new `.babelrc` file is a simplified copy of the previous preset.