[Web UI] Update browser polyfills (#1563)
## What is this change?

Upgrade the browser polyfill system to serve cacheable, conditionally-loaded bundle chunks using Modernizr for feature detection.

Replace the conditional `Promise` polyfill with a constant alias to `bluebird` better error handling and consitent general behavior in all browsers.

## Why is this change necessary?
closes #1357 

Since polyfills are only needed on some clients, it makes sense to serve them conditionally. This keeps the main bundle size to an absolute minimum.

Splitting the bundle into conditional chunks also moves more code into assets that can be cached long term. Polyfill chunks are 100% comprised of "vendor" code, and will be deterministic and cacheable across many builds and updates to the app over a long period of time. This approach improves the app load-time for all browsers.

The decision was made to force a `Promise` polyfill, even in browsers that provide a native `Promise` implementation. This is done for consistency - we are less likely to encounter promise-related bugs, and for debugability - `bluebird` offers the important `onunhandledrejection` that some browsers don't provide (this ensures our global error state is always triggered for unhandled errors) and offers better error logging in general.

## Does your change need a Changelog entry?

nah

## Do you need clarification on anything?

nope

## Were there any complications while making this change?

I haven't tested the build in IE and other browsers for which we are providing polyfills (although I have tested in Chrome by stubbing the feature detection).