Remove prop types from DrawerButton component
As we adopt new language features, prop types are continually becoming more of a nuisance than a benefit.

In this case `React.memo` returns a component that is validly defined as an object, not a function. Checking for a valid component prop with prop types is not worth the required effort.

Signed-off-by: Neal Granger <neal@nealg.com>