Display errors in status toasts (#98)
* Fix alignment of icons within toast component.

Align elements to the top edge of the toast to improve the appearance of toasts with a large amount of content.

Signed-off-by: Neal Granger <neal@nealg.com>

* Create a hook to bind a relocation component to a promise

The `usePromiseBoundComponent` hook consumes a promise and creates a unique relocation component instance and provides the consumer a callback to update the instance in response to the promise rejection or resolution.

The utility of this hook it in the way it associates a unique component instance (referenced by the unique id returned from relocation) with a particular promise instance. This allows the created component to be removed while the promise is pending - when the promise resolves or rejects the component will be recreated. If the component is not removed, it will be updated in place.

This behavior serves as an abstraction of the lifecycle logic currently implemented in each status toast component.

Signed-off-by: Neal Granger <neal@nealg.com>

* Add error statuses to existing status toasts.

Make use of the new `usePromiseBoundComponent` hook to re-implement the existing status toasts, giving the the ability to display error statuses.

Since hooks can only be used by stateless functional components - any class components which create status toasts have been migrated to functional components.

Signed-off-by: Neal Granger <neal@nealg.com>