Upgrade ApolloClient
We have been using an outdated apollo-client and a deprecated apollo-link-state since the replacement for apollo-link-state lacks support for context customization.

There doesn't appear to be any progress on that issue, so continuing to hold off waiting for a fix isn't a good idea.

I work around the problem here by creating a custom InMemoryCache subclass which includes a reset behavior that supports custom resolver default values. Apollo includes the cache instance in the non-customizable context provided to custom resolvers which is what allows to get access to functionality there.

Signed-off-by: Neal Granger <neal@nealg.com>