Add 'apollo-link-context' as a dependency
This change re-adds 'apollo-link-context' as a dependency since it's removal in `https://github.com/sensu/web/pull/144`. Without 'apollo-link-context', the interface for creating apollo links in the application is complicated because the `client` object needs to be passed to the a create functions for each link, then we need to invoke each apollo link create function when creating the apollo client. It's simpler to keep a link context around so that our apollo links can pull things off the context when needed.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>