Update `parseFilter` logic to allow for ":" values
This change updates the `parseFilter` implementation in `lib/util/filterParams.tsx` to allow filter values to inlclude ":" values. The previous implmentation used `String.split()` which discards the other split values in the string. This implementation looks for the first ":" separator, and splits the key and value, retaining the rest of the value even if it includes ":".

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>