Disable re-run action for proxy entity checks
This change disables the "re-run" button for Proxy Entity Checks and for "keepalive" checks.

Previously, the Web UI allowed users to attempt to re-run a check against a proxy entity, but it wouldn't do anyting because our application sends an empty array of subscriptions to the `executeCheck` mutation when the entity is a proxy entity.

The reason we don't preform check re-runs against proxy entities because to do so we would need to re-run the check against all its subscriptions. We believe that, as a user, clicking re-run on a check and having the check run against all subscriptions is not expected behaviour, and could lead to unwanted results.

In the future, we may implement a dialogue that would allow the user to select which subscriptions a check is run against, so that users have more control over what happens when a check is re-run. However, for the time being, we are going to remove the “re-run” button from the Check Results screen for proxy entity checks.

Signed-off-by: Wesley Yu <wesleyu89@gmail.com>